requirements.txt
colorama==0.4.6
numpy==2.3.5
Pillow==12.0.0
pywinstyles==1.8
scipy==1.16.3
sv_ttk==2.6.0
